% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curules_RP_function.R
\name{curules_rp}
\alias{curules_rp}
\title{Calcular curules de representación proporcional (RP)}
\usage{
curules_rp(datos, nulos, no_registrados, independientes)
}
\arguments{
\item{datos}{Debe ser un dataframe con tres columnas: los partidos políticos, la votación total emitida que obtuvieron, y las curules de mayoría relativa alcanzadas.}

\item{nulos}{Vector que indica qué fila(s) debe(n) ser consideradas como votos nulos}

\item{no_registrados}{Vector que indica qué fila(s) debe(n) ser consideradas como candidatos no registrados}

\item{independientes}{Vector que indica qué fila(s) debe(n) ser consideradas como candidatos independientes}
}
\description{
Esta función permite calcular las curules de representación proporcional
que corresponden a cada partido en la Cámara de Diputados de México.
}
\examples{
curules_rp(diputados_RP_2018, "VOTOS_NULOS", "CANDIDATURAS_NO_REGISTRADAS",  c("CandidaturaIndependiente_1", "CandidaturaIndependiente_2"))
}
